pyproject.toml
[project]
name = "nexushive-agent-insights"
version = "0.1.0"
description = "Enterprise-grade observability for AI agents using OpenTelemetry"
authors = [
    {name = "Aidan Pace",email = "apace@defrecord.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.10,<3.14"
dependencies = [
    "opentelemetry-api (>=1.31.1,<1.32.0)",
    "opentelemetry-sdk (>=1.31.1,<1.32.0)",
    "opentelemetry-exporter-otlp (>=1.31.1,<1.32.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
